requirements.txt
openai-whisper>=20231117
google-generativeai>=0.8.0
sounddevice>=0.4.6
numpy>=1.24.0
pyttsx3>=2.90
python-dotenv>=1.0.0
